refactor: move GitHub trust context into monochange_github
